backend/requirements.txt
# Python backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pyserial==3.5
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3

